+++ android-sdk/mixpush-android/src/main/java/cn/leancloud/AVMixPushManager.java
@@ -5,6 +5,14 @@ import android.app.Activity;
 import android.app.Application;
 import android.content.Context;
 import android.os.Build;
+import android.os.Looper;
+
+import com.huawei.agconnect.config.AGConnectServicesConfig;
+import com.huawei.hmf.tasks.OnCompleteListener;
+import com.huawei.hmf.tasks.Task;
+import com.huawei.hms.aaid.HmsInstanceId;
+import com.huawei.hms.push.HmsMessaging;
+import com.huawei.hms.support.api.push.service.HmsMsgService;
 
 import java.util.List;
 
@@ -32,7 +40,7 @@ public class AVMixPushManager {
    * 华为推送的 deviceProfile
    */
   static String hwDeviceProfile = "";
-  static Class hwPushReceiverClazz = AVHMSPushMessageReceiver.class;
+  static Class hwMessageServiceClazz = AVHMSMessageService.class;
 
   /**
    * 魅族推送的 deviceProfile
@@ -112,12 +120,12 @@ public class AVMixPushManager {
     }
 
     if (!isXiaomiPhone()) {
-      printErrorLog("register error, is not xiaomi phone!");
-      return;
+      printErrorLog("register error, current device is not a xiaomi phone!");
     }
 
     if (!checkXiaomiManifest(context)) {
-      printErrorLog("register error, mainifest is incomplete!");
+      printErrorLog("register error, mainifest is incomplete(receiver not found: "
+          + miPushReceiverClazz.getSimpleName() + ")!");
       return;
     }
 
@@ -125,7 +133,7 @@ public class AVMixPushManager {
 
     com.xiaomi.mipush.sdk.MiPushClient.registerPush(context, miAppId, miAppKey);
 
-    LOGGER.d("start register mi push");
+    LOGGER.d("finished to register mi push");
   }
 
   /**
@@ -162,23 +170,16 @@ public class AVMixPushManager {
 
   /**
    * 初始化方法，建议在 Application onCreate 里面调用
+   *
    * @param application 应用实例
    */
   public static void registerHMSPush(Application application) {
-    registerHMSPush(application, "", null);
+    registerHMSPush(application, "");
   }
 
   /**
    * 初始化方法，建议在 Application onCreate 里面调用
-   * @param application 应用实例
-   * @param customizedReceiver 自定义 receiver
-   */
-  public static void registerHMSPush(Application application, Class customizedReceiver) {
-    registerHMSPush(application, "", customizedReceiver);
-  }
-
-  /**
-   * 初始化方法，建议在 Application onCreate 里面调用
+   *
    * @param application 应用实例
    * @param profile 华为推送配置
    */
@@ -186,24 +187,18 @@ public class AVMixPushManager {
     registerHMSPush(application, profile, null);
   }
 
-  /**
-   * 初始化方法，建议在 Application onCreate 里面调用
-   * @param application 应用实例
-   * @param profile 华为推送配置
-   * @param customizedReceiver 自定义 receiver
-   */
-  public static void registerHMSPush(Application application, String profile, Class customizedReceiver) {
+  public static void registerHMSPush(Application application, String profile, Class customMessageServiceClazz) {
     if (null == application) {
       throw new IllegalArgumentException("[HMS] context cannot be null.");
-    }
-
-    if (null != customizedReceiver) {
-      hwPushReceiverClazz = customizedReceiver;
     }
 
     if (!isHuaweiPhone()) {
       printErrorLog("[HMS] register error, is not huawei phone!");
       return;
+    }
+
+    if (null != customMessageServiceClazz) {
+      hwMessageServiceClazz = customMessageServiceClazz;
     }
 
     if (!checkHuaweiManifest(application)) {
@@ -212,87 +207,100 @@ public class AVMixPushManager {
     }
 
     hwDeviceProfile = profile;
-    boolean hmsInitResult = com.huawei.android.hms.agent.HMSAgent.init(application);
-    if (!hmsInitResult) {
-      LOGGER.e("failed to init HMSAgent.");
-    }
 
     LOGGER.d("[HMS] start register HMS push");
   }
 
   /**
-   * 连接HMS SDK， 可能拉起界面(包括升级引导等)，建议在第一个界面进行连接。
-   * 此方法可以重复调用，没必要为了只调用一次做复杂处理
-   * 方法为异步调用，调用结果在主线程回调
    *  Connecting to the HMS SDK may pull up the activity (including upgrade guard, etc.), and it is
    *  recommended that you connect in the first activity.
    *  This method can be called repeatedly, and there is no need to do complex processing
    *  for only one call at a time
    *  Method is called asynchronously, and the result is invoked in the main thread callback
+   *
+   * @param activity activity
    */
   public static void connectHMS(Activity activity) {
     if (null == activity) {
       throw new IllegalArgumentException("[HMS] activity cannot be null.");
     }
-    com.huawei.android.hms.agent.HMSAgent.connect(activity,
-        new com.huawei.android.hms.agent.common.handler.ConnectHandler() {
-          @Override
-          public void onConnect(int rst) {
-            LOGGER.d("[HMS] connect end:" + rst);
-            com.huawei.android.hms.agent.HMSAgent.Push.getToken(
-                new com.huawei.android.hms.agent.push.handler.GetTokenHandler() {
-                  @Override
-                  public void onResult(int rst) {
-                    LOGGER.d("[HMS] get token: end. returnCode=" + rst);
-                  }
-                }
-            );
-          }
-        });
+    String appId = AGConnectServicesConfig.fromContext(activity).getString("client/app_id");
+    connectHMS(activity, appId);
   }
 
   /**
-   * 打开/关闭透传消息
-   *  Turn on/off notification bar messages
-   * @param enable 打开/关闭（默认为打开）
-   *                Turn ON/off
+   *  Connecting to the HMS SDK may pull up the activity (including upgrade guard, etc.), and it is
+   *  recommended that you connect in the first activity.
+   *  This method can be called repeatedly, and there is no need to do complex processing
+   *  for only one call at a time
+   *  Method is called asynchronously, and the result is invoked in the main thread callback
+   *
+   * @param activity activity
+   * @param huaweiAppId huawei app id
    */
-  public static void setHMSReceiveNormalMsg(final boolean enable) {
-    com.huawei.android.hms.agent.HMSAgent.Push.enableReceiveNormalMsg(enable,
-        new com.huawei.android.hms.agent.push.handler.EnableReceiveNormalMsgHandler() {
-          @Override
-          public void onResult(int rst) {
-            LOGGER.d("[HMS] enableReceiveNormalMsg(flag=" + enable + ") returnCode=" + rst);
+  public static void connectHMS(Activity activity, String huaweiAppId) {
+    if (null == activity) {
+      throw new IllegalArgumentException("[HMS] activity cannot be null.");
+    }
+    if (Looper.getMainLooper() == Looper.myLooper()) {
+      new Thread(new Runnable() {
+        @Override
+        public void run() {
+          try {
+            String token = HmsInstanceId.getInstance(activity).getToken(huaweiAppId, HmsMessaging.DEFAULT_TOKEN_SCOPE);
+            LOGGER.d("found HMS appId: " + huaweiAppId + ", token: " + token);
+            AVHMSMessageService.updateAVInstallation(token);
+          } catch (Exception ex) {
+            LOGGER.w("failed to get hms token. cause: " + ex.getMessage());
           }
-        });
+        }
+      }).start();
+    } else {
+      try {
+        String token = HmsInstanceId.getInstance(activity).getToken(huaweiAppId, HmsMessaging.DEFAULT_TOKEN_SCOPE);
+        LOGGER.d("found HMS appId: " + huaweiAppId + ", token: " + token);
+        AVHMSMessageService.updateAVInstallation(token);
+      } catch (Exception ex) {
+        LOGGER.w("failed to get hms token. cause: " + ex.getMessage());
+      }
+    }
+
   }
 
   /**
-   * 打开/关闭通知栏消息
-   *  Turn on/off notification bar messages
-   * @param enable 打开/关闭（默认为打开）
-   *                Turn ON/off
+   * 开启华为 HMS 推送
+   *
+   * @param context context
+   * @param callback callback function
    */
-  public static void setHMSReceiveNotifyMsg(final boolean enable) {
-    com.huawei.android.hms.agent.HMSAgent.Push.enableReceiveNotifyMsg(enable,
-        new com.huawei.android.hms.agent.push.handler.EnableReceiveNotifyMsgHandler() {
-          @Override
-          public void onResult(int rst) {
-            LOGGER.d("[HMS] enableReceiveNotifyMsg(flag=" + enable + ") returnCode=" + rst);
-          }
-        });
-  }
-
-  /**
-   * 请求push协议展示
-   *  Request Push Protocol Display
-   */
-  public static void showHMSAgreement() {
-    com.huawei.android.hms.agent.HMSAgent.Push.queryAgreement(
-        new com.huawei.android.hms.agent.push.handler.QueryAgreementHandler() {
+  public static void turnOnHMSPush(Context context, AVCallback<Void> callback) {
+    HmsMessaging.getInstance(context).turnOnPush().addOnCompleteListener(new OnCompleteListener<Void>() {
       @Override
-      public void onResult(int rst) {
-        LOGGER.d("[HMS] query agreement result: " + rst);
+      public void onComplete(Task<Void> task) {
+        if (task.isSuccessful()) {
+          callback.internalDone(null);
+        } else {
+          callback.internalDone(new AVException(task.getException()));
+        }
+      }
+    });
+  }
+
+  /**
+   * 关闭华为 HMS 推送
+   *
+   * @param context context
+   * @param callback callback function
+   */
+  public static void turnOffHMSPush(Context context, AVCallback<Void> callback) {
+    HmsMessaging.getInstance(context).turnOffPush().addOnCompleteListener(new OnCompleteListener<Void>() {
+      @Override
+      public void onComplete(Task<Void> task) {
+        if (task.isSuccessful()) {
+          callback.internalDone(null);
+        } else {
+          callback.internalDone(new AVException(task.getException()));
+        }
       }
     });
   }
@@ -387,14 +395,15 @@ public class AVMixPushManager {
 
   /**
    * 初始化方法，建议在 Application onCreate 里面调用
-   * @param application
+   * @param application application
    */
   public static boolean registerVIVOPush(Application application) {
     return AVMixPushManager.registerVIVOPush(application, "");
   }
+
   /**
    * 初始化方法，建议在 Application onCreate 里面调用
-   * @param application
+   * @param application application
    */
   public static boolean registerVIVOPush(Application application, String profile) {
     vivoDeviceProfile = profile;
@@ -596,10 +605,10 @@ public class AVMixPushManager {
   /**
    * register Oppo Push.
    *
-   * @param context
-   * @param appKey
-   * @param appSecret
-   * @param callback
+   * @param context context
+   * @param appKey oppo application key
+   * @param appSecret oppo application secret
+   * @param callback callback
    * @return
    */
   public static boolean registerOppoPush(Context context, String appKey, String appSecret,
@@ -607,9 +616,26 @@ public class AVMixPushManager {
     if (!isSupportOppoPush(context)) {
       return false;
     }
-    com.coloros.mcssdk.PushManager.getInstance().register(context, appKey, appSecret, callback);
+    com.heytap.mcssdk.PushManager.getInstance().register(context, appKey, appSecret, callback);
     return true;
   }
+
+  /**
+   * register oppo push
+   * @param context context
+   * @param appKey oppo application key
+   * @param appSecret oppo application secret
+   * @param profile profile string.
+   * @param callback callback.
+   * @return
+   */
+  public static boolean registerOppoPush(Context context, String appKey, String appSecret,
+                                         String profile,
+                                         AVOPPOPushAdapter callback) {
+    oppoDeviceProfile = profile;
+    return registerOppoPush(context, appKey, appSecret, callback);
+  }
+
 
   /**
    * judgement if support oppo push or not.
@@ -618,21 +644,21 @@ public class AVMixPushManager {
    * @return
    */
   public static boolean isSupportOppoPush(Context context) {
-    return com.coloros.mcssdk.PushManager.isSupportPush(context);
+    return com.heytap.mcssdk.PushManager.isSupportPush(context);
   }
 
   /**
    * pause oppo push
    */
   public static void pauseOppoPush() {
-    com.coloros.mcssdk.PushManager.getInstance().pausePush();
+    com.heytap.mcssdk.PushManager.getInstance().pausePush();
   }
 
   /**
    * resume oppo push
    */
   public static void resumeOppoPush() {
-    com.coloros.mcssdk.PushManager.getInstance().resumePush();
+    com.heytap.mcssdk.PushManager.getInstance().resumePush();
   }
 
   /**
@@ -645,15 +671,8 @@ public class AVMixPushManager {
    */
   public static void setOppoPushTime(List<Integer> weekDays, int startHour, int startMinute,
                                      int endHour, int endMinute) {
-    com.coloros.mcssdk.PushManager.getInstance().setPushTime(weekDays, startHour, startMinute,
+    com.heytap.mcssdk.PushManager.getInstance().setPushTime(weekDays, startHour, startMinute,
         endHour, endMinute);
-  }
-
-  /**
-   * retrieve oppo push time.
-   */
-  public static void getOppoPushTime() {
-    com.coloros.mcssdk.PushManager.getInstance().getPushTime();
   }
 
   /**
@@ -661,7 +680,7 @@ public class AVMixPushManager {
    * @param aliases
    */
   public static void setOppoAliases(List<String> aliases) {
-    com.coloros.mcssdk.PushManager.getInstance().setAliases(aliases);
+    com.heytap.mcssdk.PushManager.getInstance().setAliases(aliases);
   }
 
   /**
@@ -669,14 +688,14 @@ public class AVMixPushManager {
    * @param alias
    */
   public static void unsetOppoAlias(String alias) {
-    com.coloros.mcssdk.PushManager.getInstance().unsetAlias(alias);
+    com.heytap.mcssdk.PushManager.getInstance().unsetAlias(alias);
   }
 
   /**
    * get oppo aliases.
    */
   public static void getOppoAliases() {
-    com.coloros.mcssdk.PushManager.getInstance().getAliases();
+    com.heytap.mcssdk.PushManager.getInstance().getAliases();
   }
 
   /**
@@ -684,7 +703,7 @@ public class AVMixPushManager {
    * @param account
    */
   public static void setOppoUserAccount(String account) {
-    com.coloros.mcssdk.PushManager.getInstance().setUserAccount(account);
+    com.heytap.mcssdk.PushManager.getInstance().setUserAccount(account);
   }
 
   /**
@@ -692,14 +711,14 @@ public class AVMixPushManager {
    * @param accounts
    */
   public static void unsetOppoUserAccouts(List<String> accounts) {
-    com.coloros.mcssdk.PushManager.getInstance().unsetUserAccounts(accounts);
+    com.heytap.mcssdk.PushManager.getInstance().unsetUserAccounts(accounts);
   }
 
   /**
    * get oppo push accounts.
    */
   public static void getOppoUserAccounts() {
-    com.coloros.mcssdk.PushManager.getInstance().getUserAccounts();
+    com.heytap.mcssdk.PushManager.getInstance().getUserAccounts();
   }
 
   /**
@@ -707,7 +726,7 @@ public class AVMixPushManager {
    * @param tags
    */
   public static void setOppoTags(List<String> tags) {
-    com.coloros.mcssdk.PushManager.getInstance().setTags(tags);
+    com.heytap.mcssdk.PushManager.getInstance().setTags(tags);
   }
 
   /**
@@ -715,28 +734,28 @@ public class AVMixPushManager {
    * @param tags
    */
   public static void unsetOppoTags(List<String> tags) {
-    com.coloros.mcssdk.PushManager.getInstance().unsetTags(tags);
+    com.heytap.mcssdk.PushManager.getInstance().unsetTags(tags);
   }
 
   /**
    * retrieve oppo push tags.
    */
   public static void getOppoTags() {
-    com.coloros.mcssdk.PushManager.getInstance().getTags();
+    com.heytap.mcssdk.PushManager.getInstance().getTags();
   }
 
   /**
    * get oppo push status
    */
   public static void getOppoPushStatus() {
-    com.coloros.mcssdk.PushManager.getInstance().getPushStatus();
+    com.heytap.mcssdk.PushManager.getInstance().getPushStatus();
   }
 
   /**
    * get oppo notification status.
    */
   public static void getOppoNotificationStatus() {
-    com.coloros.mcssdk.PushManager.getInstance().getNotificationStatus();
+    com.heytap.mcssdk.PushManager.getInstance().getNotificationStatus();
   }
 
   /**
@@ -792,8 +811,10 @@ public class AVMixPushManager {
           && AVManifestUtils.checkPermission(context, android.Manifest.permission.ACCESS_NETWORK_STATE)
           && AVManifestUtils.checkPermission(context, android.Manifest.permission.ACCESS_WIFI_STATE)
           && AVManifestUtils.checkPermission(context, android.Manifest.permission.READ_PHONE_STATE)
-          && AVManifestUtils.checkReceiver(context, hwPushReceiverClazz);
+          && AVManifestUtils.checkService(context, HmsMsgService.class)
+          && AVManifestUtils.checkService(context, hwMessageServiceClazz);
     } catch (Exception e) {
+      LOGGER.d(e.getMessage());
     }
     return result;
   }
@@ -807,6 +828,7 @@ public class AVMixPushManager {
           && AVManifestUtils.checkPermission(context, Manifest.permission.WRITE_EXTERNAL_STORAGE)
           && AVManifestUtils.checkReceiver(context, flymePushReceiverClazz);
     } catch (Exception e) {
+      LOGGER.d(e.getMessage());
     }
     return result;
   }