+++ core/src/main/java/cn/leancloud/core/LoggingInterceptor.java
@@ -23,22 +23,21 @@ public class LoggingInterceptor implements Interceptor {
     StringBuilder sb = new StringBuilder();
     sb.append(String.format(CURL_COMMAND, method));
     for (String name : headers.names()) {
-//      if (RequestPaddingInterceptor.HEADER_KEY_LC_APPID.equals(name)) {
-//        sb.append(String.format(CURL_HEADER_FORMAT, name, "{your_app_id}"));
-//        continue;
-//      }
-      if (RequestPaddingInterceptor.HEADER_KEY_LC_APPKEY.equals(name)) {
-        sb.append(String.format(CURL_HEADER_FORMAT, name, "{your_app_key}"));
-        continue;
-      }
+      if (!AVOSCloud.printAllHeaders) {
+        if (RequestPaddingInterceptor.HEADER_KEY_LC_APPKEY.equals(name)) {
+          sb.append(String.format(CURL_HEADER_FORMAT, name, "{your_app_key}"));
+          continue;
+        }
 
-      if (RequestPaddingInterceptor.HEADER_KEY_LC_SESSIONTOKEN.equals(name)) {
-        sb.append(String.format(CURL_HEADER_FORMAT, name, "{your_session}"));
-        continue;
-      }
-      if (RequestPaddingInterceptor.HEADER_KEY_LC_SIGN.equals(name)) {
-        sb.append(String.format(CURL_HEADER_FORMAT, RequestPaddingInterceptor.HEADER_KEY_LC_SIGN, "{your_sign}"));
-        continue;
+        if (RequestPaddingInterceptor.HEADER_KEY_LC_SESSIONTOKEN.equals(name)) {
+          sb.append(String.format(CURL_HEADER_FORMAT, name, "{your_session}"));
+          continue;
+        }
+
+        if (RequestPaddingInterceptor.HEADER_KEY_LC_SIGN.equals(name)) {
+          sb.append(String.format(CURL_HEADER_FORMAT, RequestPaddingInterceptor.HEADER_KEY_LC_SIGN, "{your_sign}"));
+          continue;
+        }
       }
       sb.append(String.format(CURL_HEADER_FORMAT, name, headers.get(name)));
     }