+++ CHANGELOG.md
@@ -2,6 +2,36 @@
 
 Following is change logs for recently release versions, you can refer to [releases page](https://github.com/leancloud/java-unified-sdk/releases) for more details.
 
+## 6.4.4 release
+
+#### Break changes
+- None
+
+#### New features
+- AVMixPushManager support to connect HMS with specified huawei appId:
+```java
+AVMixPushManager.connectHMS(Activity activity, String huaweiAppId);
+```
+
+- AVMixPushManager support to register Oppo with specified profile:
+```java
+AVMixPushManager.registerOppoPush(Context context, String appKey, String appSecret,
+                                  String profile, AVOPPOPushAdapter callback)
+```
+
+- add `AVOSCloud.enablePrintAllHeaders4Debug(boolean flag)` method to print the whole request(usable for android securely initialization).
+
+#### Optimization and fixed bugs
+- upgrade mixpush with latest third-party SDKs:
+  - xiaomi: 3.6.9 -> 3.7.5
+  - huawei(HMS): 2.6.3.306 -> 4.0.2.300
+  - oppo: 1.0.1 -> 2.0.2
+  - vivo: 2.3.4 -> 2.9.0.0
+
+- remove duplicated value within AVInstallation.channels
+
+- support customized subclass of AVIMTypedMessage in Kotlin.
+
 ## 6.4.3 release
 
 #### Break changes
@@ -12,7 +42,6 @@ Following is change logs for recently release versions, you can refer to [releas
 
 #### Optimization and fixed bugs
 - fixed: can't create conversation with signature.
-
 
 ## 6.4.2 release